// types/states.ts
export interface LinkItem {
    name: string;
    href: string;
  }
  
  export interface StateSection {
    title: string;
    icon: string;
    items: LinkItem[];
  }
  
  export interface StateInfo {
    name: string;
    abbreviation: string;
    capital: string;
    population: string;
    website: string;
    sections: StateSection[];
  }export interface LinkItem {
    name: string;
    href: string;
  }
  
  export interface StateSection {
    title: string;
    icon: string;
    items: LinkItem[];
  }
  
  export interface StateInfo {
    name: string;
    abbreviation: string;
    capital: string;
    population: string;
    website: string;
    sections: StateSection[];
  }